lint by forge fmt

+++ src/interfaces/IWeth.sol
@@ -5,7 +5,7 @@ pragma solidity ^0.8.13;
 interface IWETH {
     function deposit() external payable;
 
-    function transfer(address to, uint value) external returns (bool);
+    function transfer(address to, uint256 value) external returns (bool);
 
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
 }
